+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Markers of lightness are only merged when labelLevel == 0.
     - When merge happens, the marker shows the middle value of lightness.
 - Remove rare fantom empty marker space on scrollable legend.
+- Same legend with different measure series are not interpolated. 
 
 ## [0.15.0] - 2024-10-28